requirements.txt
pywebview>=4.4.1
pywinusb>=0.4.2
keyboard>=0.13.5
pyautogui>=0.9.54
pystray>=0.19.4
Pillow>=10.0.0
pywin32>=306
